requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
Django==4.0.5
django-cors-headers==3.13.0
filelock==3.7.1
huggingface-hub==0.8.1
idna==3.3
joblib==1.1.0
numpy==1.23.1
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
PyYAML==6.0
regex==2022.7.9
requests==2.28.1
sacremoses==0.0.53
six==1.16.0
sqlparse==0.4.2
tokenizers==0.12.1
torch==1.12.0+cu116
torchaudio==0.12.0+cu116
torchvision==0.13.0+cu116
tqdm==4.64.0
transformers==4.18.0
typing-extensions==4.3.0
tzdata==2022.1
urllib3==1.26.10
